need to figure out how to structure the library to handle singular/plural nouns

+++ src/main/java/org/tsd/tsdbot/haloapi/model/stats/Enemy.java
@@ -6,6 +6,6 @@ public class Enemy {
     long BaseId;
 
     // The attachments (variants) for the enemy.
-    int[] Attachments;
+    long[] Attachments;
 
 }
